refactor(component): simplify icon component
