+++ src/lib/components/icon/index.ts
@@ -1,2 +1,1 @@
 export { default as Icon } from "./icon.svelte";
-export { default as SVG } from "./svg.svelte";
